Add mle_fit function
Begin to define the mle fit in terms of a function that defines and
entire graph and then exectues the graph in a while loop in the session.
The function needs to be rewritten so that it is largely predefiend in
the graph. At the moment too much memory needs to be allocated
repetitivley in the loop.